test-requirements.txt
tox>=2.9.1
pytest>=3.4.0
pytest-tornado>=0.4.5
pytest-asyncio>=0.8.0
coverage>=4.5.0
